+++ src/Services/Catalog/OrderSphere.Catalog.Infrastructure/GlobalUsings.cs
@@ -1,5 +1,7 @@
 global using Microsoft.EntityFrameworkCore;
 global using Microsoft.EntityFrameworkCore.Design;
 global using Microsoft.EntityFrameworkCore.Metadata.Builders;
+global using OrderSphere.BuildingBlocks.Blob;
+global using OrderSphere.BuildingBlocks.Infrastructure.Blob;
 global using OrderSphere.Catalog.Application.Abstractions;
 global using OrderSphere.Catalog.Domain.Entities;