+++ tests/OrderSphere.Catalog.Tests/GlobalUsings.cs
@@ -1,10 +1,11 @@
 global using FluentAssertions;
 global using NSubstitute;
+global using OrderSphere.BuildingBlocks.Blob;
+global using OrderSphere.BuildingBlocks.Infrastructure.Blob;
 global using OrderSphere.BuildingBlocks.StronglyTypedIds;
 global using OrderSphere.BuildingBlocks.ValueObjects;
 global using OrderSphere.Catalog.Application.Abstractions;
 global using OrderSphere.Catalog.Domain.Entities;
 global using OrderSphere.Catalog.Domain.Errors;
-global using OrderSphere.Catalog.Infrastructure.Blob;
 global using OrderSphere.Catalog.Infrastructure.Search;
 global using Xunit;